+++ src/maths/fft/fftlib.h
@@ -1,14 +1,14 @@
 #define MYRECIPLN2	1.442695040888963407359924681001892137426	// 1.0/log(2)
 
 /* some useful conversions between a number and its power of 2 */
-#define LOG2(a)	(MYRECIPLN2*log(a))	// floating point logarithm base 2
+#define LOG2(a)	(MYRECIPLN2*log(a))	// doubleing point logarithm base 2
 #define POW2(m) ((unsigned long) 1 << (m))	// integer power of 2 for m<32
 
 /*******************************************************************
 lower level fft stuff called by routines in fftext.c and fft2d.c
 *******************************************************************/
 
-void fftCosInit(long M, float *Utbl);
+void fftCosInit(long M, double *Utbl);
 /* Compute Utbl, the cosine table for ffts	*/
 /* of size (pow(2,M)/4 +1)	*/
 /* INPUTS */
@@ -24,7 +24,7 @@ void fftBRInit(long M, short *BRLow);
 /* OUTPUTS */
 /* *BRLow = bit reversed counter table	*/
 
-void ffts1(float *ioptr, long M, long Rows, float *Utbl, short *BRLow);
+void ffts1(double *ioptr, long M, long Rows, double *Utbl, short *BRLow);
 /* Compute in-place complex fft on the rows of the input array	*/
 /* INPUTS */
 /* *ioptr = input data array	*/
@@ -35,7 +35,7 @@ void ffts1(float *ioptr, long M, long Rows, float *Utbl, short *BRLow);
 /* OUTPUTS */
 /* *ioptr = output data array	*/
 
-void iffts1(float *ioptr, long M, long Rows, float *Utbl, short *BRLow);
+void iffts1(double *ioptr, long M, long Rows, double *Utbl, short *BRLow);
 /* Compute in-place inverse complex fft on the rows of the input array	*/
 /* INPUTS */
 /* *ioptr = input data array	*/
@@ -46,7 +46,7 @@ void iffts1(float *ioptr, long M, long Rows, float *Utbl, short *BRLow);
 /* OUTPUTS */
 /* *ioptr = output data array	*/
 
-void rffts1(float *ioptr, long M, long Rows, float *Utbl, short *BRLow);
+void rffts1(double *ioptr, long M, long Rows, double *Utbl, short *BRLow);
 /* Compute in-place real fft on the rows of the input array	*/
 /* The result is the complex spectra of the positive frequencies */
 /* except the location for the first complex number contains the real */
@@ -62,7 +62,7 @@ void rffts1(float *ioptr, long M, long Rows, float *Utbl, short *BRLow);
 /* Re(x[0]), Re(x[N/2]), Re(x[1]), Im(x[1]), Re(x[2]), Im(x[2]), ... Re(x[N/2-1]), Im(x[N/2-1]). */
 
 
-void riffts1(float *ioptr, long M, long Rows, float *Utbl, short *BRLow);
+void riffts1(double *ioptr, long M, long Rows, double *Utbl, short *BRLow);
 /* Compute in-place real ifft on the rows of the input array	*/
 /* data order as from rffts1 */
 /* INPUTS */
